refactor custom documents tests, update recordings
